+++ src/features/index/index.js
@@ -1,5 +1,5 @@
 export { Feed } from "../feed/Feed";
 export {Explore} from "../explore/Explore"; 
-export {UserProfilePage} from "../userProfilePage/UserProfilePage";
+export {ProfilePage} from "../ProfilePage/ProfilePage";
 export {BookMarks} from "../bookmarks/BookMarks";
 export {Login, SignUp, } from "../auth";
