+++ src/routes/AllRoutes/Routes.js
@@ -1,5 +1,5 @@
 import {Routes, Route} from "react-router-dom";
-import { BookMarks, Explore, Feed, UserProfilePage, Login, SignUp} from "../../features/index/index";
+import { BookMarks, Explore, Feed, ProfilePage, Login, SignUp} from "../../features/index/index";
 import { AuthRequiresRoutes } from "../AuthRequires/AuthRquiresRoutes";
 import { Template } from "../TemplateRoute/Template";
 
@@ -30,7 +30,7 @@ export const AllRoutes = () => {
             <Route path="/userProfile" element={
                 <AuthRequiresRoutes>
                     <Template>
-                        <UserProfilePage />
+                        <ProfilePage />
                     </Template>
                 </AuthRequiresRoutes>
             } />
